+++ frontend/src/api/authentication.ts
@@ -2,7 +2,7 @@ import { END_POINTS } from '@/routes';
 import type { LoginRequest, SignupRequest } from '@/types';
 import { MemberInfo } from '@/types';
 
-import { apiClient } from './ApiClient';
+import { apiClient } from './config';
 import { customFetch } from './customFetch';
 
 const API_URL = process.env.REACT_APP_API_URL || 'https://default-url.com';
