+++ frontend/src/api/config.ts
@@ -1,0 +1,10 @@
+import { ApiClient } from './ApiClient';
+
+const API_URL = process.env.REACT_APP_API_URL ?? 'https://default-url.com';
+
+const httpHeader = {
+  'Content-Type': 'application/json',
+};
+const httpCredentials = 'include';
+
+export const apiClient = new ApiClient(API_URL, httpHeader, httpCredentials);
